Changing double to BigDecimal

+++ xchange-core/src/main/java/org/knowm/xchange/utils/PriceUtils.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class PriceUtils {
 
-    public static BigDecimal tickRounding(BigDecimal price, double tickSize){
-        return BigDecimal.valueOf(price.doubleValue() - price.doubleValue() % tickSize + ((price.doubleValue() % tickSize < tickSize / 2) ? 0.0 : tickSize)).stripTrailingZeros();
+    public static BigDecimal tickRounding(BigDecimal price, BigDecimal tickSize){
+        return BigDecimal.valueOf(price.doubleValue() - price.doubleValue() % tickSize.doubleValue() + ((price.doubleValue() % tickSize.doubleValue() < tickSize.doubleValue() / 2) ? 0.0 : tickSize.doubleValue())).stripTrailingZeros();
     }
 }
